pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fengwen2"
version = "1.0.0"
description = "Astrology Fortune Backend Service"
authors = [
    {name = "fengwen2", email = "info@fengwen2.com"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sqlalchemy>=1.4.0",
    "python-multipart>=0.0.5",
    "python-dotenv>=0.20.0",
    "httpx>=0.24.0",
    "zhdate",
    "tencentcloud-sdk-python",
    "pydantic[email]>=1.10.0",
    "jinja2>=3.0.0",
    "playwright>=1.48.0",
    "requests>=2.32.4",
    "beautifulsoup4>=4.13.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.html", "*.css", "*.js"]

[project.urls]
Homepage = "https://github.com/fengwen2/astrology-backend"
Repository = "https://github.com/fengwen2/astrology-backend"

[project.scripts]
fengwen2 = "fengwen2.main:main"
